% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biased_rw.R
\name{biased_rw}
\alias{biased_rw}
\title{biased random walk function}
\usage{
biased_rw(
  n = 1,
  data,
  xy = NULL,
  coa = NULL,
  dir = NULL,
  buffer = NULL,
  rho,
  a,
  b
)
}
\description{
utility function not to be called by user
}
