% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{growth (energetics) function}
\usage{
growth(w, ts, s)
}
\arguments{
\item{w}{mass at previous time step in g}

\item{ts}{water temperature at previous time step in C}

\item{s}{active swim speed at previous time step in km/h}
}
\description{
utility function not to be called by user
}
