% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsim.R
\name{mapsim}
\alias{mapsim}
\title{mapsim}
\usage{
mapsim(
  x,
  data = NULL,
  xlim = NULL,
  ylim = NULL,
  res = 5,
  esrf = FALSE,
  rec = TRUE,
  track = TRUE,
  tcol = "black",
  last = FALSE,
  alpha = 0.5,
  lwd = 0.25,
  size = 0.2,
  col = "red",
  pal = "Blues 3",
  crs = "+proj=laea +lat_0=41 +lon_0=-71 +units=km +datum=WGS84",
  ...
)
}
\arguments{
\item{x}{a fitted object of class simsmolt}

\item{data}{data object created by sim_setup}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{res}{downsampling factor to product a plot faster (default = 5, full-resolution = 0)}

\item{rec}{should receiver locations be displayed (logical)}

\item{alpha}{translucence for smolt track(s)}

\item{lwd}{width of smolt track(s)}

\item{size}{of smolt track end point(s)}

\item{col}{colour for receiver locations}
}
\description{
Produce a simple map of simulated track(s) with or without receivers
}
