% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R, R/moveKcam.R
\name{moveKcam}
\alias{moveKcam}
\title{random walk movement kernel for Campbellton River kelts}
\usage{
moveKcam(data, xy = NULL, mpar, i, s, ts, w)

moveKcam(data, xy = NULL, mpar, i, s, ts, w)
}
\description{
utility function not to be called by user

utility function not to be called by user
}
