% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdet.r
\name{pdet}
\alias{pdet}
\title{Utility fn to calculate distances b/w transmitter and reciever}
\usage{
pdet(trs = NULL, rec = NULL, b = NULL)
}
\arguments{
\item{trs}{- a tibble of transmission locations (in m) with: id, x, y, et and line id}

\item{rec}{- a data.frame of receiver locations (in m) with: x, y, z}

\item{b}{- a vector of two parameters (intercept & slope) for logistic detection range function}
}
\description{
assumes detection is a logistic fn of distance in 3-D
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
