% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_drifter.R
\name{sim_drifter}
\alias{sim_drifter}
\title{simulate drifter advection by surface currents from user-specified start location(s)}
\usage{
sim_drifter(id = 1, data = NULL, mpar = sim_par(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual drifter)}

\item{data}{- a list of required data from \code{sim_setup}}

\item{mpar}{- simulation control parameters supplied as a list, see details}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates drifter advection
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
