% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_growth (original).R, R/sim_growth.R
\name{sim_growth}
\alias{sim_growth}
\title{simulate movements with or without current advection, using inputs from sim_setup}
\usage{
sim_growth(id = 1, N = 2760, data = NULL, mpar = list(), pb = TRUE)

sim_growth(id = 1, N = 2760, data = NULL, mpar = list(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual animal)}

\item{N}{- number of 1-h time steps to simulate}

\item{data}{- a list of required data from \code{presim}}

\item{mpar}{- simulation control parameters supplied as a list, see details}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates movement

simulates movement
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}

Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
