% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_move.R
\name{sim_move}
\alias{sim_move}
\title{simulate movements & survival, using inputs from presim}
\usage{
sim_move(N = 1800, tag = c(150, 240), coa = c(900, 950),
  chgpt = c(800, 835), data = NULL, param = list(a = 2, b = 0.864,
  rho = 0.8, ntries = 100, surv = 0.9936))
}
\arguments{
\item{N}{- number of 1-h time steps to simulate}

\item{tag}{- start location(s) of simulated animals}

\item{coa}{- optional Centre-Of-Attraction location(s) to provide movement bias(es)}

\item{data}{- a list of required data from \code{presim}}

\item{params}{- movement & (otpional) survival parameters supplied as a list, see details}
}
\description{
simulates, movement & detection of acoustically-tagged animals
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
