% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(bathy = file.path("..", "simdata", "prob_xy.grd"), uv = NULL,
  sst = NULL, rec = TRUE)
}
\arguments{
\item{bathy}{- required bathymetry layer to define simulation domain}

\item{uv}{- optional current layers, supplied as u and v components, velocity must be in m/s}

\item{sst}{- optional sea surface temperature layer(s)}

\item{rec}{- optional acoustic receiver locations}

\item{tag}{- start location(s) of simulated animals}

\item{coa}{- optional Centre-Of-Attraction location(s) to provide movement bias(es)}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
