% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_brw.R
\name{temp_brw}
\alias{temp_brw}
\title{temperature-biased random walk function}
\usage{
temp_brw(
  n = 1,
  i = NULL,
  mpar = NULL,
  d1 = NULL,
  data,
  xy = NULL,
  ts = NULL,
  ts.rng = NULL,
  b
)
}
\description{
utility function not to be called by user
}
