% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simsmolt.R
\name{plot.simsmolt}
\alias{plot.simsmolt}
\title{plot}
\usage{
\method{plot}{simsmolt}(s, data, xlim = NULL, ylim = NULL,
  ca = FALSE, raster = TRUE, alpha = 0.25, lwd = 0.2, ...)
}
\arguments{
\item{s}{a fitted object of class simsmolt}

\item{data}{data object created by sim_setup}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{ca}{plot centres of attraction}
}
\description{
Map simulated track
}
