% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simsmolt.R
\name{plot.simsmolt}
\alias{plot.simsmolt}
\title{plot}
\usage{
\method{plot}{simsmolt}(
  x,
  data = NULL,
  xlim = NULL,
  ylim = NULL,
  raster = "ts",
  res = 5,
  esrf = TRUE,
  layer = NULL,
  rec = FALSE,
  track = TRUE,
  last = FALSE,
  alpha = 0.5,
  lwd = 0.2,
  size = 0.2,
  col = c("salmon", "red", "dodgerblue"),
  pal = "Cividis",
  maxp = 50000,
  landfill = grey(0.6),
  crs = "+proj=laea +lat_0=41 +lon_0=-71 +units=km +datum=WGS84",
  ...
)
}
\arguments{
\item{x}{a fitted object of class simsmolt}

\item{data}{data object created by sim_setup}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{raster}{select a raster (from data) to use as background (bathy, uv, or NULL)}

\item{res}{downsampling factor to product a plot faster (default = 5, full-resolution = 0)}

\item{rec}{should receiver locations be displayed (logical)}

\item{track}{should smolt track(s) be displayed (logical)}

\item{last}{display position at end of simulation only (logical; if T then track forced to be F)}

\item{alpha}{translucence for smolt track(s)}

\item{lwd}{width of smolt track(s)}

\item{size}{of smolt track end point(s)}

\item{col}{colour for receiver locations}
}
\description{
Map simulated track
}
