% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simsmolt.R
\name{plot.simsmolt}
\alias{plot.simsmolt}
\title{plot}
\usage{
\method{plot}{simsmolt}(
  s,
  data,
  xlim = NULL,
  ylim = NULL,
  raster = "ts",
  bathy = FALSE,
  fast = TRUE,
  layer = NULL,
  rec = FALSE,
  track = TRUE,
  alpha = 0.9,
  lwd = 0.2,
  size = 0.2,
  col = "red",
  pal = "Cividis",
  maxp = 50000,
  crs = "+proj=laea +lat_0=41 +lon_0=-71 +units=km +ellps=WGS84",
  ...
)
}
\arguments{
\item{s}{a fitted object of class simsmolt}

\item{data}{data object created by sim_setup}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}

\item{raster}{select a raster (from data) to use as background (bathy, uv, or NULL)}

\item{fast}{generate plot quickly (poorer resolution; default = TRUE)}

\item{rec}{should receiver locations be displayed (logical)}

\item{track}{should smolt track(s) be displayed (logical)}

\item{alpha}{translucence for smolt track(s)}

\item{lwd}{width of smolt track(s)}

\item{size}{of smolt track end point(s)}

\item{col}{colour for receiver locations}
}
\description{
Map simulated track
}
