% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simsmolt.R
\name{plot.simsmolt}
\alias{plot.simsmolt}
\title{plot}
\usage{
\method{plot}{simsmolt}(s, xlim = NULL, ylim = NULL, ca = TRUE, ...)
}
\arguments{
\item{s}{a fitted object of class simsmolt}

\item{xlim}{plot x limits}

\item{ylim}{plot y limits}
}
\description{
Map simulated track
}
