% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_detect.r
\name{sim_detect}
\alias{sim_detect}
\title{simulate acoustic transmissions & detection, using sim_move & presim output}
\usage{
sim_detect(s, delay = c(20, 60), burst = 5)
}
\arguments{
\item{s}{- a simsmolt class list containing output from sim_setup and sim_move}

\item{delay}{- min & max time intervals (s) between transmissions}

\item{burst}{- duration of each transmission (s)}
}
\description{
simulates transmissions & detections along simulated track segments within a defined range of acoustic array(s)
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
