% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_move.R
\name{sim_move}
\alias{sim_move}
\title{simulate movements & survival, using inputs from presim}
\usage{
sim_move(id = 1, N = 2760, data = NULL, mpar = list(), pb = TRUE)
}
\arguments{
\item{id}{- identifier for simulation run (individual animal)}

\item{N}{- number of 1-h time steps to simulate}

\item{data}{- a list of required data from \code{presim}}

\item{mpar}{- simulation control parameters supplied as a list, see details}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates, movement & detection of acoustically-tagged animals
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
