% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_move.R
\name{sim_move}
\alias{sim_move}
\title{simulate movements & survival, using inputs from presim}
\usage{
sim_move(N = 1800, tag = c(712, 761), data = NULL, mpar = list())
}
\arguments{
\item{N}{- number of 1-h time steps to simulate}

\item{tag}{- start location(s) of simulated animals}

\item{data}{- a list of required data from \code{presim}}

\item{mpar}{- movement & (otpional) survival parameters supplied as a list, see details}

\item{coa}{- optional Centre-Of-Attraction location(s) to provide movement bias(es)}
}
\description{
simulates, movement & detection of acoustically-tagged animals
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
