% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(
<<<<<<< HEAD
  config = file.path("..", "simdata", "config.R"),
  rec = "none",
  rspace = NULL,
  rnum = NULL,
  ocean = "doy",
  doy.rng = NULL,
  uv = FALSE,
  crs = NULL
=======
  bathy = file.path("..", "simdata", "bathy_xy.grd"),
  land = file.path("..", "simdata", "d2landc_xy.grd"),
  land.dir = file.path("..", "simdata", "land_dirc.grd"),
  b900.dist = file.path("..", "simdata", "b900_dist.grd"),
  b900.dir = file.path("..", "simdata", "b900_dir.grd"),
  u = file.path("..", "simdata", "riops_u.grd"),
  v = file.path("..", "simdata", "riops_v.grd"),
  sst = NULL,
  rec = "lines",
  rspace = 10,
  rnum = 135,
  uv = FALSE
>>>>>>> master
)
}
\arguments{
\item{config}{- path to config.R script containing file.paths for required & optional environmental layers (see Details)}

\item{rec}{- optional acoustic receiver locations (default is no receivers)}

\item{rspace}{- nominal spacing (km) between receivers}

\item{rnum}{- the number of receivers to be used on grid arrays (to approx. match to num used on lines)}

\item{ocean}{- temporal form of environmental data layers: "cl" = climatological; "m" = monthly means; "wk" = weekly means}

\item{doy.rng}{- rnage of days of year to be loaded from disk as a rasterStack (default = 1:365)}

\item{uv}{- are current velocity data to be loaded from disk}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
