% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(
  config = file.path("..", "simdata", "config.R"),
  rec = "none",
  ocean = "cl"
)
}
\arguments{
\item{config}{- path to config.R script containing file.paths for required & optional environmental layers (see Details)}

\item{rec}{- optional acoustic receiver locations (default is no receivers)}

\item{tag}{- start location(s) of simulated animals}

\item{coa}{- optional Centre-Of-Attraction location(s) to provide movement bias(es)}

\item{u}{- optional zonal current layers, velocity must be in m/s}

\item{v}{- optional meridional current layers, velocity must be in m/s}

\item{ts}{- optional sea surface temperature layer(s)}

\item{rspace}{- nominal spacing (km) between receivers}

\item{rnum}{- the number of receivers to be used on grid arrays (to approx. match to num used on lines)}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
