% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(
  config = file.path("/", "Users", "jonsen", "OneDrive - Macquarie University",
    "collab", "otn", "esrf", "fn", "config.R")
)
}
\arguments{
\item{config}{- path to config.R script containing file.paths for required & optional environmental layers (see Details)}

\item{rec}{- optional acoustic receiver locations (default is no receivers)}

\item{rspace}{- nominal spacing (km) between receivers}

\item{rnum}{- the number of receivers to be used on grid arrays (to approx. match to num used on lines)}

\item{ocean}{- temporal form of environmental data layers: "cl" = climatological; "doy" = daily means; FALSE = no env't data}

\item{doy.rng}{- rnage of days of year to be loaded from disk as a rasterStack (default = 1:365)}

\item{uv}{- are current velocity data to be loaded from disk}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
