% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.r
\name{sim_setup}
\alias{sim_setup}
\title{Pre-simulation setup}
\usage{
sim_setup(bathy = file.path("..", "simdata", "bathy.nc"),
  land = file.path("..", "simdata", "d2land_xy.grd"),
  land.dir = file.path("..", "simdata", "land_dir2.grd"),
  b700.dist = file.path("..", "simdata", "b700_dist.grd"),
  b700.dir = file.path("..", "simdata", "b700_dir.grd"), uv = NULL,
  sst = NULL, rec = TRUE)
}
\arguments{
\item{bathy}{- required bathymetry file}

\item{land}{- required distance to land to define simulation domain}

\item{land.dir}{- required direction to nearest land}

\item{b700.dist}{- required distance to 700 m isobath}

\item{b700.dir}{-  required direction to 700 m isobath}

\item{uv}{- optional current layers, supplied as u and v components, velocity must be in m/s}

\item{sst}{- optional sea surface temperature layer(s)}

\item{rec}{- optional acoustic receiver locations}

\item{tag}{- start location(s) of simulated animals}

\item{coa}{- optional Centre-Of-Attraction location(s) to provide movement bias(es)}
}
\description{
load required rasters, receiver locations
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
